+++ packages/db/drizzle/meta/_journal.json
@@ -159,8 +159,8 @@
     {
       "idx": 22,
       "version": "7",
-      "when": 1729749846860,
-      "tag": "0022_aspiring_xorn",
+      "when": 1729806557821,
+      "tag": "0022_clean_wrecking_crew",
       "breakpoints": true
     }
   ]
