+++ packages/db/src/schema/target.ts
@@ -194,7 +194,6 @@ export function targetMatchesMetadata(
 export const targetRelationshipType = pgEnum("target_relationship_type", [
   "associated_with",
   "depends_on",
-  "extends",
 ]);
 
 export const targetRelationship = pgTable(
